+++ tests/com/kaylerrenslow/armaDialogCreator/gui/fx/main/treeview/EditorComponentTreeViewTest.java
@@ -43,7 +43,7 @@ public class EditorComponentTreeViewTest extends Application {
 	private static class TestArmaControlClass extends ArmaControl {
 
 		public TestArmaControlClass() {
-			super("", new ArmaControlSpecRequirement(), new ArmaResolution(ScreenDimension.D960, ArmaUIScale.DEFAULT), RendererLookup.TEST, new Env() {
+			super("", ArmaControlSpecRequirement.TRIVIAL, new ArmaResolution(ScreenDimension.D960, ArmaUIScale.DEFAULT), RendererLookup.TEST, new Env() {
 				@Override
 				public @Nullable Value getValue(String identifier) {
 					return null;